requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
jinja2>=3.1.2

